Update Person
Ahora Person tienen position y no fighter
