+++ person/Person.java
@@ -1,14 +1,18 @@
 package person;
 
+import map.Position2D;
+
 public abstract class Person {
     private String name;
+    private Position2D position;
 
     /**
      * Constructor of Person. It assigns a name to the Person object
      * @param name name of the player/enemy.
      */
-    public Person (String name){
+    public Person (String name, Position2D position){
         this.name=name;
+        this.position=position;
     }
 
     public String getName(){
@@ -24,5 +28,13 @@ public abstract class Person {
 
         if (this.getName().equals(other.getName())) return true;
         return false;
+    }
+
+    public Position2D getPosition() {
+        return position;
+    }
+
+    public void setPosition(Position2D pos){
+        position=pos;
     }
 }
